+++ lib/github/model/tag.rb
@@ -4,7 +4,7 @@ module Kraken
   module GitHub
     class Tag
       attr_accessor :name
-      def initialize(name)
+      def initialize(name:)
         @name = name
       end
     end