+++ lib/kubernetes/model/pod.rb
@@ -4,7 +4,7 @@ module Kraken
   module Kubernetes
     class Pod
       attr_accessor :name, :status
-      def initialize(name = nil, status = nil)
+      def initialize(name: nil, status: nil)
         @name = name
         @status = status
       end